# src/ai_travel_planner/models/create_plan_request.py
# coding: utf-8

"""
    AI Travel Planner API

    No description provided (generated by Openapi Generator https://github.com/openapitools/openapi-generator)

    The version of the OpenAPI document: 1.0.0
    Generated by OpenAPI Generator (https://openapi-generator.tech)

    Do not edit the class manually.
"""  # noqa: E501


from __future__ import annotations
import pprint
import re  # noqa: F401
import json




from datetime import date
from pydantic import BaseModel, ConfigDict, Field, StrictStr
from typing import Any, ClassVar, Dict, List, Optional
from typing_extensions import Annotated
from ai_travel_planner.models.budget import Budget
try:
    from typing import Self
except ImportError:
    from typing_extensions import Self

class CreatePlanRequest(BaseModel):
    """
    CreatePlanRequest
    """ # noqa: E501
    destination: StrictStr
    start_date: date
    end_date: date
    budget: Budget
    interests: Optional[List[StrictStr]] = None
    travelers: Annotated[int, Field(strict=True, ge=1)]
    __properties: ClassVar[List[str]] = ["destination", "start_date", "end_date", "budget", "interests", "travelers"]

    model_config = {
        "populate_by_name": True,
        "validate_assignment": True,
        "protected_namespaces": (),
    }


    def to_str(self) -> str:
        """Returns the string representation of the model using alias"""
        return pprint.pformat(self.model_dump(by_alias=True))

    def to_json(self) -> str:
        """Returns the JSON representation of the model using alias"""
        # TODO: pydantic v2: use .model_dump_json(by_alias=True, exclude_unset=True) instead
        return json.dumps(self.to_dict())

    @classmethod
    def from_json(cls, json_str: str) -> Self:
        """Create an instance of CreatePlanRequest from a JSON string"""
        return cls.from_dict(json.loads(json_str))

    def to_dict(self) -> Dict[str, Any]:
        """Return the dictionary representation of the model using alias.

        This has the following differences from calling pydantic's
        `self.model_dump(by_alias=True)`:

        * `None` is only added to the output dict for nullable fields that
          were set at model initialization. Other fields with value `None`
          are ignored.
        """
        _dict = self.model_dump(
            by_alias=True,
            exclude={
            },
            exclude_none=True,
        )
        # override the default output from pydantic by calling `to_dict()` of budget
        if self.budget:
            _dict['budget'] = self.budget.to_dict()
        return _dict

    @classmethod
    def from_dict(cls, obj: Dict) -> Self:
        """Create an instance of CreatePlanRequest from a dict"""
        if obj is None:
            return None

        if not isinstance(obj, dict):
            return cls.model_validate(obj)

        _obj = cls.model_validate({
            "destination": obj.get("destination"),
            "start_date": obj.get("start_date"),
            "end_date": obj.get("end_date"),
            "budget": Budget.from_dict(obj.get("budget")) if obj.get("budget") is not None else None,
            "interests": obj.get("interests"),
            "travelers": obj.get("travelers")
        })
        return _obj