# src/ai_travel_planner/models/plan_status.py
# coding: utf-8

"""
    AI Travel Planner API

    No description provided (generated by Openapi Generator https://github.com/openapitools/openapi-generator)

    The version of the OpenAPI document: 1.0.0
    Generated by OpenAPI Generator (https://openapi-generator.tech)

    Do not edit the class manually.
"""  # noqa: E501


from __future__ import annotations
import json
import pprint
import re  # noqa: F401
from enum import Enum



try:
    from typing import Self
except ImportError:
    from typing_extensions import Self


class PlanStatus(str, Enum):
    """
    PlanStatus
    """

    """
    allowed enum values
    """
    PENDING = 'pending'
    RESEARCHING = 'researching'
    PLANNING = 'planning'
    AWAITING_REVIEW = 'awaiting_review'
    REVISING = 'revising'
    APPROVED = 'approved'
    FAILED = 'failed'

    @classmethod
    def from_json(cls, json_str: str) -> Self:
        """Create an instance of PlanStatus from a JSON string"""
        return cls(json.loads(json_str))