# src/ai_travel_planner/models/workflow_stage.py
# coding: utf-8

"""
    AI Travel Planner API

    No description provided (generated by Openapi Generator https://github.com/openapitools/openapi-generator)

    The version of the OpenAPI document: 1.0.0
    Generated by OpenAPI Generator (https://openapi-generator.tech)

    Do not edit the class manually.
"""  # noqa: E501


from __future__ import annotations
import json
import pprint
import re  # noqa: F401
from enum import Enum



try:
    from typing import Self
except ImportError:
    from typing_extensions import Self


class WorkflowStage(str, Enum):
    """
    WorkflowStage
    """

    """
    allowed enum values
    """
    INTAKE = 'intake'
    RESEARCH = 'research'
    ITINERARY_GENERATION = 'itinerary_generation'
    HITL_REVIEW = 'hitl_review'
    REVISION = 'revision'
    FINALIZED = 'finalized'

    @classmethod
    def from_json(cls, json_str: str) -> Self:
        """Create an instance of WorkflowStage from a JSON string"""
        return cls(json.loads(json_str))